mock children calc functions
