+++ packages/react-components/src/components/ResponsiveHorizontalGallery.tsx
@@ -4,8 +4,8 @@
 import { IStyle, mergeStyles } from '@fluentui/react';
 import React, { useRef } from 'react';
 import { HorizontalGallery, HorizontalGalleryStyles } from './HorizontalGallery';
-import { _convertRemToPx as convertRemToPx } from '@internal/acs-ui-common';
 import { _useContainerWidth } from './utils/responsive';
+import { calculateHorizontalChildrenPerPage } from './VideoGallery/utils/OverflowGalleryUtils';
 
 /**
  * Wrapped HorizontalGallery that adjusts the number of items per page based on the
@@ -24,18 +24,10 @@ export const ResponsiveHorizontalGallery = (props: {
   const containerRef = useRef<HTMLDivElement>(null);
   const containerWidth = _useContainerWidth(containerRef);
 
-  const leftPadding = containerRef.current
-    ? !isNaN(parseFloat(getComputedStyle(containerRef.current).paddingLeft))
-      ? parseFloat(getComputedStyle(containerRef.current).paddingLeft)
-      : 0
-    : 0;
-  const rightPadding = containerRef.current
-    ? !isNaN(parseFloat(getComputedStyle(containerRef.current).paddingRight))
-      ? parseFloat(getComputedStyle(containerRef.current).paddingRight)
-      : 0
-    : 0;
+  const leftPadding = containerRef.current ? parseFloat(getComputedStyle(containerRef.current).paddingLeft) : 0;
+  const rightPadding = containerRef.current ? parseFloat(getComputedStyle(containerRef.current).paddingRight) : 0;
 
-  const childrenPerPage = calculateChildrenPerPage({
+  const childrenPerPage = calculateHorizontalChildrenPerPage({
     numberOfChildren: React.Children.count(props.children),
     containerWidth: (containerWidth ?? 0) - leftPadding - rightPadding,
     childWidthRem,
@@ -54,51 +46,4 @@ export const ResponsiveHorizontalGallery = (props: {
       </HorizontalGallery>
     </div>
   );
-};
-
-/**
- * Helper function to calculate children per page for HorizontalGallery based on width of container, child, buttons, and
- * gaps in between
- */
-const calculateChildrenPerPage = (args: {
-  numberOfChildren: number;
-  containerWidth: number;
-  childWidthRem: number;
-  gapWidthRem: number;
-  buttonWidthRem: number;
-}): number => {
-  const { numberOfChildren, containerWidth, buttonWidthRem, childWidthRem, gapWidthRem } = args;
-
-  const childWidth = convertRemToPx(childWidthRem);
-  const gapWidth = convertRemToPx(gapWidthRem);
-
-  /** First check how many children can fit in containerWidth.
-   *    __________________________________
-   *   |                ||                |
-   *   |                ||                |
-   *   |________________||________________|
-   *   <-----------containerWidth--------->
-   *  containerWidth = n * childWidth + (n - 1) * gapWidth. Isolate n and take the floor.
-   */
-  const numberOfChildrenInContainer = Math.floor((containerWidth + gapWidth) / (childWidth + gapWidth));
-  // If all children fit then return numberOfChildrenInContainer
-  if (numberOfChildren <= numberOfChildrenInContainer) {
-    return numberOfChildrenInContainer;
-  }
-
-  const buttonWidth = convertRemToPx(buttonWidthRem);
-
-  /** We know we need to paginate. So we need to subtract the buttonWidth twice and gapWidth twice from
-   * containerWidth to compute childrenSpace
-   *   <-----------containerWidth--------->
-   *    __________________________________
-   *   | ||             ||             || |
-   *   |<||             ||             ||>|
-   *   |_||_____________||_____________||_|
-   *       <-------childrenSpace------>
-   */
-  const childrenSpace = containerWidth - 2 * buttonWidth - 2 * gapWidth;
-  // Now that we have childrenSpace width we can figure out how many children can fit in childrenSpace.
-  // childrenSpace = n * childWidth + (n - 1) * gapWidth. Isolate n and take the floor.
-  return Math.max(Math.floor((childrenSpace + gapWidth) / (childWidth + gapWidth)), 1);
 };
